anchor version 0.22.1 -> 0.24.2, stake program structure
