+++ Anchor.toml
@@ -1,3 +1,6 @@
+anchor_version = "0.24.2"
+solana_version = "1.9.12"
+
 [features]
 seeds = false
 [programs.mainnet]
