+++ programs/stake/Cargo.toml
@@ -16,8 +16,8 @@ cpi = ["no-entrypoint"]
 default = []
 
 [dependencies]
-anchor-lang = { version = "0.22.1", features = ["init-if-needed"]}
-anchor-spl = "0.22.1"
+anchor-lang = ">=0.22"
+anchor-spl = ">=0.22"
 spl-token = { version = "3.3.0", features = ["no-entrypoint"] }
 num-traits = "0.2"
 ve_honey = { path = "../ve_honey", features = ["cpi"], version = "0.1" }
