+++ programs/stake/src/constants.rs
@@ -1,0 +1,7 @@
+pub const POOL_INFO_SEED: &str = "PoolInfo";
+pub const POOL_USER_SEED: &str = "PoolUser";
+pub const TOKEN_VAULT_SEED: &str = "TokenVault";
+pub const AUTHORITY_SEED: &str = "VaultAuthority";
+pub const STAKE_POOL_VERSION: u8 = 1;
+pub const CLAIM_PERIOD_UNIT: i64 = 86_400;
+pub const CLAIM_MAX_COUNT: u8 = 21;
