+++ programs/stake/src/context.rs
@@ -1,0 +1,19 @@
+pub mod claim;
+pub mod deposit;
+pub mod init;
+pub mod init_user;
+pub mod modify_params;
+pub mod reclaim_mint_authority;
+pub mod set_mint_authority;
+pub mod set_owner;
+pub mod stake;
+
+pub use claim::*;
+pub use deposit::*;
+pub use init::*;
+pub use init_user::*;
+pub use modify_params::*;
+pub use reclaim_mint_authority::*;
+pub use set_mint_authority::*;
+pub use set_owner::*;
+pub use stake::*;
