+++ programs/ve_honey/Cargo.toml
@@ -16,8 +16,8 @@ cpi = ["no-entrypoint"]
 default = []
 
 [dependencies]
-anchor-lang = { version = "0.22.1", features = ["init-if-needed"] }
-anchor-spl = "0.22.1"
+anchor-lang = { version = ">=0.22", features = ["init-if-needed"] }
+anchor-spl = ">=0.22"
 bitflags = "1.3.2"
 govern = { version = "^0.5", features = ["cpi"] }
 mpl-token-metadata = { version = "1.3.4", features = ["no-entrypoint"] }
